fix JsonObject.getInt() - getting long

+++ org.zenframework.z8.server/src/main/java/org/zenframework/z8/server/base/json/parser/JsonObject.java
@@ -105,7 +105,7 @@ public class JsonObject extends OBJECT {
 	}
 
 	public integer z8_getInt(string name) {
-		return new integer(object.getInt(name));
+		return new integer(object.getLong(name));
 	}
 
 	public integer z8_getInt(JsonPath.CLASS<? extends JsonPath> path) {
